+++ components/debug/stresstest/src/main/java/com/flipperdevices/debug/stresstest/api/StressTestApiImpl.kt
@@ -1,16 +1,16 @@
 package com.flipperdevices.debug.stresstest.api
 
+import androidx.compose.runtime.Composable
 import com.flipperdevices.core.di.AppGraph
 import com.flipperdevices.debug.api.StressTestApi
-import com.flipperdevices.debug.stresstest.fragments.StressTestFragment
-import com.github.terrakok.cicerone.Screen
-import com.github.terrakok.cicerone.androidx.FragmentScreen
+import com.flipperdevices.debug.stresstest.composable.ComposableStressTestScreen
 import com.squareup.anvil.annotations.ContributesBinding
 import javax.inject.Inject
 
 @ContributesBinding(AppGraph::class)
 class StressTestApiImpl @Inject constructor() : StressTestApi {
-    override fun getStressTestScreen(): Screen {
-        return FragmentScreen { StressTestFragment() }
+    @Composable
+    override fun StressTestScreen() {
+        ComposableStressTestScreen()
     }
 }