+++ components/screenstreaming/impl/src/main/java/com/flipperdevices/screenstreaming/impl/viewmodel/ScreenStreamingViewModel.kt
@@ -19,7 +19,6 @@ import javax.inject.Inject
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
 import kotlinx.coroutines.launch
@@ -94,5 +93,19 @@ class ScreenStreamingViewModel : LifecycleViewModel() {
                 guiStopScreenStreamRequest = stopScreenStreamRequest {}
             }.wrapToRequest()
         ).launchIn(viewModelScope)
+    }
+
+    fun enableStreaming() {
+        streamingState.compareAndSet(
+            expect = StreamingState.DISABLED,
+            update = StreamingState.ENABLED
+        )
+    }
+
+    fun disableStreaming() {
+        streamingState.compareAndSet(
+            expect = StreamingState.ENABLED,
+            update = StreamingState.DISABLED
+        )
     }
 }