+++ components/settings/impl/src/main/java/com/flipperdevices/settings/impl/viewmodels/ExperimentalViewModel.kt
@@ -6,7 +6,6 @@ import com.flipperdevices.core.di.ComponentHolder
 import com.flipperdevices.core.preference.pb.Settings
 import com.flipperdevices.core.ui.LifecycleViewModel
 import com.flipperdevices.filemanager.api.navigation.FileManagerScreenProvider
-import com.flipperdevices.screenstreaming.api.ScreenStreamingApi
 import com.flipperdevices.settings.impl.di.SettingsComponent
 import com.github.terrakok.cicerone.Router
 import javax.inject.Inject
@@ -19,9 +18,6 @@ class ExperimentalViewModel : LifecycleViewModel() {
     lateinit var fileManager: FileManagerScreenProvider
 
     @Inject
-    lateinit var screenStreamingApi: ScreenStreamingApi
-
-    @Inject
     lateinit var dataStoreSetting: DataStore<Settings>
 
     init {
@@ -30,9 +26,5 @@ class ExperimentalViewModel : LifecycleViewModel() {
 
     fun onOpenFileManager(router: Router) {
         router.navigateTo(fileManager.fileManager())
-    }
-
-    fun onScreenStreaming(router: Router) {
-        router.navigateTo(screenStreamingApi.provideScreen())
     }
 }