+++ components/settings/impl/src/main/java/com/flipperdevices/settings/impl/viewmodels/SettingsViewModel.kt
@@ -7,6 +7,8 @@ import androidx.lifecycle.viewModelScope
 import com.flipperdevices.core.di.ComponentHolder
 import com.flipperdevices.core.navigation.global.CiceroneGlobal
 import com.flipperdevices.core.preference.pb.Settings
+import com.flipperdevices.debug.api.StressTestApi
+import com.flipperdevices.screenstreaming.api.ScreenStreamingApi
 import com.flipperdevices.settings.impl.di.SettingsComponent
 import com.flipperdevices.shake2report.api.Shake2ReportApi
 import javax.inject.Inject
@@ -24,6 +26,12 @@ class SettingsViewModel : ViewModel() {
 
     @Inject
     lateinit var shakeToReportApi: Shake2ReportApi
+
+    @Inject
+    lateinit var screenStreamingApi: ScreenStreamingApi
+
+    @Inject
+    lateinit var stressTestApi: StressTestApi
 
     init {
         ComponentHolder.component<SettingsComponent>().inject(this)