+++ CHANGELOG.md
@@ -29,5 +29,6 @@
 - [Feature] Add disconnect button and improve reconnect
 - [Feature] Now you can delete or restore all keys from trash
 - [Feature] Add recover support to deleted keys (on keyscreen)
+- [Feature] Compose Navigation in "Options" 
 - [CI] Now we check that CHANGELOG.md changed
 - [CI] JVM protobuf library