+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
-# 1.1.2 - In progress
+# 1.1.3 - In progress
+
+# 1.1.2B
 
 - [BUGFIX] Fix crash with custom firmware version
 - [BUGFIX] Fix crash when key file not exist
@@ -33,6 +35,6 @@
 - [Feature] Add disconnect button and improve reconnect
 - [Feature] Now you can delete or restore all keys from trash
 - [Feature] Add recover support to deleted keys (on keyscreen)
-- [Feature] Compose Navigation in "Options" 
+- [Feature] Compose Navigation in "Options"
 - [CI] Now we check that CHANGELOG.md changed
 - [CI] JVM protobuf library